feat: add controllers and endpoints for Cart and Product entities
- Implemented Cart and Product controllers with RESTful endpoints.
- Added handlers for creating, retrieving, and deleting carts and products.
- Integrated use cases with HTTP layer for cart and product operations.
- Enhanced error handling for not found and already existing resources.
- Updated router to register new cart and product routes.
- Added in-memory repositories and mocks to support controller unit tests.
- Improved test utilities for HTTP request simulation.

+++ internal/core/usecases/delete_cart.go
@@ -1,6 +1,9 @@
 package usecases
 
-import "github.com/natasha-m-oliveira/clean-architecture-go/internal/core/repositories"
+import (
+	"github.com/natasha-m-oliveira/clean-architecture-go/internal/core/errors"
+	"github.com/natasha-m-oliveira/clean-architecture-go/internal/core/repositories"
+)
 
 type (
 	DeleteCartUseCase interface {
@@ -23,12 +26,14 @@ func NewDeleteCartUseCase(cartsRepository repositories.CartsRepository) DeleteCa
 }
 
 func (uc deleteCartUseCase) Execute(request DeleteCartRequest) error {
-	_, err := uc.cartsRepository.FindById(request.Id)
+	cart, err := uc.cartsRepository.FindById(request.Id)
 	if err != nil {
 		return err
 	}
 
-	err = uc.cartsRepository.DeleteById(request.Id)
+	if cart == nil {
+		return errors.NewCartNotFound()
+	}
 
-	return err
+	return uc.cartsRepository.DeleteById(request.Id)
 }
